refactor(parser): extract types to pkg kongstate.
In more detail:
- create package KongState
- move all the types (KongState as well as wrappers for kong.* types
such as Service, Route, Upstream, etc) to package kongstate (to
individual .go and _test.go files)
- make functions that mutate the state of a wrapper (or a wrapped kong
entity) a member function of the wrapper object (for Route, Service, Upstream).
This required changes in tests:
    - to operate on the wrapper (e.g. kongstate.Route)
    - refactored assertions that calling methods on nil shall not panic
instead of the wrapped type (kong.Route)
- create package 'util' for helpers with no side effects
