+++ internal/ingress/controller/parser/parser_test.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/kong/go-kong/kong"
 	"github.com/kong/kubernetes-ingress-controller/internal/ingress/annotations"
+	"github.com/kong/kubernetes-ingress-controller/internal/ingress/controller/parser/kongstate"
 	"github.com/kong/kubernetes-ingress-controller/internal/ingress/store"
 	"github.com/kong/kubernetes-ingress-controller/internal/ingress/utils"
 	configurationv1 "github.com/kong/kubernetes-ingress-controller/pkg/apis/configuration/v1"
@@ -594,12 +595,12 @@ func TestSecretConfigurationPlugin(t *testing.T) {
 		assert.Nil(err)
 		assert.NotNil(state)
 		for _, testcase := range references {
-			config, err := secretToConfiguration(store, *testcase, "default")
+			config, err := kongstate.SecretToConfiguration(store, *testcase, "default")
 			assert.NotEmpty(config)
 			assert.Nil(err)
 		}
 		for _, testcase := range badReferences {
-			config, err := secretToConfiguration(store, *testcase, "default")
+			config, err := kongstate.SecretToConfiguration(store, *testcase, "default")
 			assert.Empty(config)
 			assert.NotEmpty(err)
 		}
@@ -2367,7 +2368,7 @@ func TestParserSecret(t *testing.T) {
 			return strings.Compare(*state.Certificates[0].SNIs[i],
 				*state.Certificates[0].SNIs[j]) > 0
 		})
-		assert.Equal(Certificate{
+		assert.Equal(kongstate.Certificate{
 			Certificate: kong.Certificate{
 				ID:   kong.String("3e8edeca-7d23-4e02-84c9-437d11b746a6"),
 				Cert: kong.String(tlsPairs[0].Cert),
@@ -3038,7 +3039,7 @@ func TestParseIngressRules(t *testing.T) {
 		parsedInfo := parseIngressRules(logrus.New(), []*networking.Ingress{},
 			[]*configurationv1beta1.TCPIngress{})
 		assert.Equal(ingressRules{
-			ServiceNameToServices: make(map[string]Service),
+			ServiceNameToServices: make(map[string]kongstate.Service),
 			SecretNameToSNIs:      make(map[string][]string),
 		}, parsedInfo)
 	})
@@ -3046,7 +3047,7 @@ func TestParseIngressRules(t *testing.T) {
 		parsedInfo := parseIngressRules(logrus.New(), []*networking.Ingress{},
 			[]*configurationv1beta1.TCPIngress{tcpIngressList[0]})
 		assert.Equal(ingressRules{
-			ServiceNameToServices: make(map[string]Service),
+			ServiceNameToServices: make(map[string]kongstate.Service),
 			SecretNameToSNIs:      make(map[string][]string),
 		}, parsedInfo)
 	})
@@ -3331,14 +3332,14 @@ func TestParseKnativeIngressRules(t *testing.T) {
 	}
 	t.Run("no ingress returns empty info", func(t *testing.T) {
 		parsedInfo := parseKnativeIngressRules([]*knative.Ingress{})
-		assert.Equal(map[string]Service{}, parsedInfo.ServiceNameToServices)
+		assert.Equal(map[string]kongstate.Service{}, parsedInfo.ServiceNameToServices)
 		assert.Equal(newSecretNameToSNIs(), parsedInfo.SecretNameToSNIs)
 	})
 	t.Run("empty ingress returns empty info", func(t *testing.T) {
 		parsedInfo := parseKnativeIngressRules([]*knative.Ingress{
 			ingressList[0],
 		})
-		assert.Equal(map[string]Service{}, parsedInfo.ServiceNameToServices)
+		assert.Equal(map[string]kongstate.Service{}, parsedInfo.ServiceNameToServices)
 		assert.Equal(newSecretNameToSNIs(), parsedInfo.SecretNameToSNIs)
 	})
 	t.Run("basic knative Ingress resource is parsed", func(t *testing.T) {
@@ -3424,837 +3425,6 @@ func TestParseKnativeIngressRules(t *testing.T) {
 		}, svc.Plugins[0])
 
 		assert.Equal(newSecretNameToSNIs(), parsedInfo.SecretNameToSNIs)
-	})
-}
-
-func TestOverrideService(t *testing.T) {
-	assert := assert.New(t)
-
-	testTable := []struct {
-		inService      Service
-		inKongIngresss configurationv1.KongIngress
-		outService     Service
-		inAnnotation   map[string]string
-	}{
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("http"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("http"),
-					Path:     kong.String("/"),
-				},
-			},
-			map[string]string{},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("http"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{
-					Protocol: kong.String("https"),
-				},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("https"),
-					Path:     kong.String("/"),
-				},
-			},
-			map[string]string{},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("http"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{
-					Retries: kong.Int(0),
-				},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("http"),
-					Path:     kong.String("/"),
-					Retries:  kong.Int(0),
-				},
-			},
-			map[string]string{},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("http"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{
-					Path: kong.String("/new-path"),
-				},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("http"),
-					Path:     kong.String("/new-path"),
-				},
-			},
-			map[string]string{},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("http"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{
-					Retries: kong.Int(1),
-				},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("http"),
-					Path:     kong.String("/"),
-					Retries:  kong.Int(1),
-				},
-			},
-			map[string]string{},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("http"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{
-					ConnectTimeout: kong.Int(100),
-					ReadTimeout:    kong.Int(100),
-					WriteTimeout:   kong.Int(100),
-				},
-			},
-			Service{
-				Service: kong.Service{
-					Host:           kong.String("foo.com"),
-					Port:           kong.Int(80),
-					Name:           kong.String("foo"),
-					Protocol:       kong.String("http"),
-					Path:           kong.String("/"),
-					ConnectTimeout: kong.Int(100),
-					ReadTimeout:    kong.Int(100),
-					WriteTimeout:   kong.Int(100),
-				},
-			},
-			map[string]string{},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("grpc"),
-					Path:     nil,
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{
-					Protocol: kong.String("grpc"),
-				},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("grpc"),
-					Path:     nil,
-				},
-			},
-			map[string]string{},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("https"),
-					Path:     nil,
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{
-					Protocol: kong.String("grpcs"),
-				},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("grpcs"),
-					Path:     nil,
-				},
-			},
-			map[string]string{},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("https"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{
-					Protocol: kong.String("grpcs"),
-				},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("grpcs"),
-					Path:     nil,
-				},
-			},
-			map[string]string{"configuration.konghq.com/protocol": "grpcs"},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("https"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{
-					Protocol: kong.String("grpcs"),
-				},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("grpc"),
-					Path:     nil,
-				},
-			},
-			map[string]string{"configuration.konghq.com/protocol": "grpc"},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("https"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("grpcs"),
-					Path:     nil,
-				},
-			},
-			map[string]string{"configuration.konghq.com/protocol": "grpcs"},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("https"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{
-					Protocol: kong.String("grpcs"),
-				},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("https"),
-					Path:     kong.String("/"),
-				},
-			},
-			map[string]string{"configuration.konghq.com/protocol": "https"},
-		},
-		{
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("https"),
-					Path:     kong.String("/"),
-				},
-			},
-			configurationv1.KongIngress{
-				Proxy: &kong.Service{},
-			},
-			Service{
-				Service: kong.Service{
-					Host:     kong.String("foo.com"),
-					Port:     kong.Int(80),
-					Name:     kong.String("foo"),
-					Protocol: kong.String("https"),
-					Path:     kong.String("/"),
-				},
-			},
-			map[string]string{"configuration.konghq.com/protocol": "https"},
-		},
-	}
-
-	for _, testcase := range testTable {
-		overrideService(&testcase.inService, &testcase.inKongIngresss, testcase.inAnnotation)
-		assert.Equal(testcase.inService, testcase.outService)
-	}
-
-	assert.NotPanics(func() {
-		overrideService(nil, nil, nil)
-	})
-}
-
-func TestOverrideRoute(t *testing.T) {
-	assert := assert.New(t)
-
-	testTable := []struct {
-		inRoute        Route
-		inKongIngresss configurationv1.KongIngress
-		outRoute       Route
-	}{
-		{
-			Route{
-				Route: kong.Route{
-					Hosts: kong.StringSlice("foo.com", "bar.com"),
-				},
-			},
-			configurationv1.KongIngress{},
-			Route{
-				Route: kong.Route{
-					Hosts: kong.StringSlice("foo.com", "bar.com"),
-				},
-			},
-		},
-		{
-			Route{
-				Route: kong.Route{
-					Hosts: kong.StringSlice("foo.com", "bar.com"),
-				},
-			},
-			configurationv1.KongIngress{
-				Route: &kong.Route{
-					Methods: kong.StringSlice("GET", "POST"),
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Hosts:   kong.StringSlice("foo.com", "bar.com"),
-					Methods: kong.StringSlice("GET", "POST"),
-				},
-			},
-		},
-		{
-			Route{
-				Route: kong.Route{
-					Hosts: kong.StringSlice("foo.com", "bar.com"),
-				},
-			},
-			configurationv1.KongIngress{
-				Route: &kong.Route{
-					Methods: kong.StringSlice("GET   ", "post"),
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Hosts:   kong.StringSlice("foo.com", "bar.com"),
-					Methods: kong.StringSlice("GET", "POST"),
-				},
-			},
-		},
-		{
-			Route{
-				Route: kong.Route{
-					Hosts: kong.StringSlice("foo.com", "bar.com"),
-				},
-			},
-			configurationv1.KongIngress{
-				Route: &kong.Route{
-					Methods: kong.StringSlice("GET", "-1"),
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Hosts: kong.StringSlice("foo.com", "bar.com"),
-				},
-			},
-		},
-		{
-			Route{
-				Route: kong.Route{
-					Hosts: kong.StringSlice("foo.com", "bar.com"),
-				},
-			},
-			configurationv1.KongIngress{
-				Route: &kong.Route{
-					HTTPSRedirectStatusCode: kong.Int(302),
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Hosts:                   kong.StringSlice("foo.com", "bar.com"),
-					HTTPSRedirectStatusCode: kong.Int(302),
-				},
-			},
-		},
-		{
-			Route{
-				Route: kong.Route{
-					Hosts:        kong.StringSlice("foo.com", "bar.com"),
-					PreserveHost: kong.Bool(true),
-					StripPath:    kong.Bool(true),
-				},
-			},
-			configurationv1.KongIngress{
-				Route: &kong.Route{
-					Protocols:     kong.StringSlice("http"),
-					PreserveHost:  kong.Bool(false),
-					StripPath:     kong.Bool(false),
-					RegexPriority: kong.Int(10),
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Hosts:         kong.StringSlice("foo.com", "bar.com"),
-					Protocols:     kong.StringSlice("http"),
-					PreserveHost:  kong.Bool(false),
-					StripPath:     kong.Bool(false),
-					RegexPriority: kong.Int(10),
-				},
-			},
-		},
-		{
-			Route{
-				Route: kong.Route{
-					Hosts:     kong.StringSlice("foo.com"),
-					Protocols: kong.StringSlice("http", "https"),
-				},
-			},
-			configurationv1.KongIngress{
-				Route: &kong.Route{
-					Headers: map[string][]string{
-						"foo-header": {"bar-value"},
-					},
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Hosts:     kong.StringSlice("foo.com"),
-					Protocols: kong.StringSlice("http", "https"),
-					Headers: map[string][]string{
-						"foo-header": {"bar-value"},
-					},
-				},
-			},
-		},
-		{
-			Route{
-				Route: kong.Route{
-					Hosts: kong.StringSlice("foo.com"),
-				},
-			},
-			configurationv1.KongIngress{
-				Route: &kong.Route{
-					Protocols: kong.StringSlice("grpc", "grpcs"),
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Hosts:     kong.StringSlice("foo.com"),
-					Protocols: kong.StringSlice("grpc", "grpcs"),
-					StripPath: nil,
-				},
-			},
-		},
-		{
-			Route{
-				Route: kong.Route{
-					Hosts: kong.StringSlice("foo.com"),
-				},
-			},
-			configurationv1.KongIngress{
-				Route: &kong.Route{
-					PathHandling: kong.String("v1"),
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Hosts:        kong.StringSlice("foo.com"),
-					PathHandling: kong.String("v1"),
-				},
-			},
-		},
-	}
-
-	for _, testcase := range testTable {
-		overrideRoute(logrus.New(), &testcase.inRoute, &testcase.inKongIngresss)
-		assert.Equal(testcase.inRoute, testcase.outRoute)
-	}
-
-	assert.NotPanics(func() {
-		overrideRoute(logrus.New(), nil, nil)
-	})
-}
-
-func TestOverrideRoutePriority(t *testing.T) {
-	assert := assert.New(t)
-	var route Route
-	route = Route{
-		Route: kong.Route{
-			Hosts: kong.StringSlice("foo.com", "bar.com"),
-		},
-	}
-	kongIngress := configurationv1.KongIngress{
-		Route: &kong.Route{
-			Hosts: kong.StringSlice("foo.com", "bar.com"),
-		},
-	}
-
-	netIngress := networking.Ingress{
-		ObjectMeta: metav1.ObjectMeta{
-			Annotations: map[string]string{
-				"configuration.konghq.com/protocols": "grpc,grpcs",
-			},
-		},
-	}
-
-	route = Route{
-		Route: kong.Route{
-			Hosts: kong.StringSlice("foo.com", "bar.com"),
-		},
-		Ingress: netIngress,
-	}
-	overrideRoute(logrus.New(), &route, &kongIngress)
-	assert.Equal(route.Hosts, kong.StringSlice("foo.com", "bar.com"))
-	assert.Equal(route.Protocols, kong.StringSlice("grpc", "grpcs"))
-}
-
-func TestOverrideRouteByKongIngress(t *testing.T) {
-	assert := assert.New(t)
-	route := Route{
-		Route: kong.Route{
-			Hosts: kong.StringSlice("foo.com", "bar.com"),
-		},
-	}
-	kongIngress := configurationv1.KongIngress{
-		Route: &kong.Route{
-			Hosts: kong.StringSlice("foo.com", "bar.com"),
-		},
-	}
-
-	overrideRouteByKongIngress(logrus.New(), &route, &kongIngress)
-	assert.Equal(route.Hosts, kong.StringSlice("foo.com", "bar.com"))
-	assert.NotPanics(func() {
-		overrideRoute(logrus.New(), nil, nil)
-	})
-}
-func TestOverrideRouteByAnnotation(t *testing.T) {
-	assert := assert.New(t)
-	var route Route
-	route = Route{
-		Route: kong.Route{
-			Hosts: kong.StringSlice("foo.com", "bar.com"),
-		},
-	}
-
-	netIngress := networking.Ingress{
-		ObjectMeta: metav1.ObjectMeta{
-			Annotations: map[string]string{
-				"configuration.konghq.com/protocols": "grpc,grpcs",
-			},
-		},
-	}
-
-	route = Route{
-		Route: kong.Route{
-			Hosts: kong.StringSlice("foo.com", "bar.com"),
-		},
-		Ingress: netIngress,
-	}
-	overrideRouteByAnnotation(logrus.New(), &route)
-	assert.Equal(route.Hosts, kong.StringSlice("foo.com", "bar.com"))
-	assert.Equal(route.Protocols, kong.StringSlice("grpc", "grpcs"))
-
-	assert.NotPanics(func() {
-		overrideRoute(logrus.New(), nil, nil)
-	})
-}
-
-func TestNormalizeProtocols(t *testing.T) {
-	assert := assert.New(t)
-	testTable := []struct {
-		inRoute  Route
-		outRoute Route
-	}{
-		{
-			Route{
-				Route: kong.Route{
-					Protocols: kong.StringSlice("grpc", "grpcs"),
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Protocols: kong.StringSlice("grpc", "grpcs"),
-				},
-			},
-		},
-		{
-			Route{
-				Route: kong.Route{
-					Protocols: kong.StringSlice("http", "https"),
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Protocols: kong.StringSlice("http", "https"),
-				},
-			},
-		},
-		{
-			Route{
-				Route: kong.Route{
-					Protocols: kong.StringSlice("grpc", "https"),
-				},
-			},
-			Route{
-				Route: kong.Route{
-					Protocols: kong.StringSlice("http", "https"),
-				},
-			},
-		},
-	}
-
-	for _, testcase := range testTable {
-		normalizeProtocols(&testcase.inRoute)
-		assert.Equal(testcase.inRoute.Protocols, testcase.outRoute.Protocols)
-	}
-
-	assert.NotPanics(func() {
-		overrideUpstream(nil, nil, make(map[string]string))
-	})
-}
-
-func TestValidateProtocol(t *testing.T) {
-	assert := assert.New(t)
-	testTable := []struct {
-		input  string
-		result bool
-	}{
-		{"http", true},
-		{"https", true},
-		{"grpc", true},
-		{"grpcs", true},
-		{"grcpsfdsafdsfafdshttp", false},
-	}
-	for _, testcase := range testTable {
-		isMatch := validateProtocol(testcase.input)
-		assert.Equal(isMatch, testcase.result)
-	}
-
-	assert.NotPanics(func() {
-		overrideUpstream(nil, nil, make(map[string]string))
-	})
-}
-func TestUseSSLProtocol(t *testing.T) {
-	assert := assert.New(t)
-	testTable := []struct {
-		inRoute  kong.Route
-		outRoute kong.Route
-	}{
-		{
-			kong.Route{
-				Protocols: kong.StringSlice("grpc", "grpcs"),
-			},
-			kong.Route{
-				Protocols: kong.StringSlice("grpcs"),
-			},
-		},
-		{
-			kong.Route{
-				Protocols: kong.StringSlice("http", "https"),
-			},
-			kong.Route{
-				Protocols: kong.StringSlice("https"),
-			},
-		},
-		{
-			kong.Route{
-				Protocols: kong.StringSlice("grpcs", "https"),
-			},
-
-			kong.Route{
-				Protocols: kong.StringSlice("grpcs", "https"),
-			},
-		},
-		{
-			kong.Route{
-				Protocols: kong.StringSlice("grpc", "http"),
-			},
-			kong.Route{
-				Protocols: kong.StringSlice("grpcs", "https"),
-			},
-		},
-		{
-			kong.Route{
-				Protocols: []*string{},
-			},
-			kong.Route{
-				Protocols: kong.StringSlice("https"),
-			},
-		},
-	}
-
-	for _, testcase := range testTable {
-		useSSLProtocol(&testcase.inRoute)
-		assert.Equal(testcase.inRoute.Protocols, testcase.outRoute.Protocols)
-	}
-}
-
-func TestOverrideUpstream(t *testing.T) {
-	assert := assert.New(t)
-
-	testTable := []struct {
-		inUpstream     Upstream
-		inKongIngresss configurationv1.KongIngress
-		outUpstream    Upstream
-	}{
-		{
-			Upstream{
-				Upstream: kong.Upstream{
-					Name: kong.String("foo.com"),
-				},
-			},
-			configurationv1.KongIngress{
-				Upstream: &kong.Upstream{},
-			},
-			Upstream{
-				Upstream: kong.Upstream{
-					Name: kong.String("foo.com"),
-				},
-			},
-		},
-		{
-			Upstream{
-				Upstream: kong.Upstream{
-					Name: kong.String("foo.com"),
-				},
-			},
-			configurationv1.KongIngress{
-				Upstream: &kong.Upstream{
-					Name:               kong.String("wrong.com"),
-					HashOn:             kong.String("HashOn"),
-					HashOnCookie:       kong.String("HashOnCookie"),
-					HashOnCookiePath:   kong.String("HashOnCookiePath"),
-					HashOnHeader:       kong.String("HashOnHeader"),
-					HashFallback:       kong.String("HashFallback"),
-					HashFallbackHeader: kong.String("HashFallbackHeader"),
-					Slots:              kong.Int(42),
-				},
-			},
-			Upstream{
-				Upstream: kong.Upstream{
-					Name:               kong.String("foo.com"),
-					HashOn:             kong.String("HashOn"),
-					HashOnCookie:       kong.String("HashOnCookie"),
-					HashOnCookiePath:   kong.String("HashOnCookiePath"),
-					HashOnHeader:       kong.String("HashOnHeader"),
-					HashFallback:       kong.String("HashFallback"),
-					HashFallbackHeader: kong.String("HashFallbackHeader"),
-					Slots:              kong.Int(42),
-				},
-			},
-		},
-	}
-
-	for _, testcase := range testTable {
-		overrideUpstream(&testcase.inUpstream, &testcase.inKongIngresss, make(map[string]string))
-		assert.Equal(testcase.inUpstream, testcase.outUpstream)
-	}
-
-	assert.NotPanics(func() {
-		overrideUpstream(nil, nil, make(map[string]string))
 	})
 }
 
@@ -4632,658 +3802,6 @@ func TestGetEndpoints(t *testing.T) {
 			result := getEndpoints(logrus.New(), testCase.svc, testCase.port, testCase.proto, testCase.fn)
 			if len(testCase.result) != len(result) {
 				t.Errorf("expected %v Endpoints but got %v", testCase.result, len(result))
-			}
-		})
-	}
-}
-
-func Test_getCombinations(t *testing.T) {
-	type args struct {
-		relations foreignRelations
-	}
-	tests := []struct {
-		name string
-		args args
-		want []rel
-	}{
-		{
-			name: "empty",
-			args: args{
-				relations: foreignRelations{},
-			},
-			want: nil,
-		},
-		{
-			name: "plugins on consumer only",
-			args: args{
-				relations: foreignRelations{
-					Consumer: []string{"foo", "bar"},
-				},
-			},
-			want: []rel{
-				{
-					Consumer: "foo",
-				},
-				{
-					Consumer: "bar",
-				},
-			},
-		},
-		{
-			name: "plugins on service only",
-			args: args{
-				relations: foreignRelations{
-					Service: []string{"foo", "bar"},
-				},
-			},
-			want: []rel{
-				{
-					Service: "foo",
-				},
-				{
-					Service: "bar",
-				},
-			},
-		},
-		{
-			name: "plugins on routes only",
-			args: args{
-				relations: foreignRelations{
-					Route: []string{"foo", "bar"},
-				},
-			},
-			want: []rel{
-				{
-					Route: "foo",
-				},
-				{
-					Route: "bar",
-				},
-			},
-		},
-		{
-			name: "plugins on service and routes only",
-			args: args{
-				relations: foreignRelations{
-					Route:   []string{"foo", "bar"},
-					Service: []string{"foo", "bar"},
-				},
-			},
-			want: []rel{
-				{
-					Service: "foo",
-				},
-				{
-					Service: "bar",
-				},
-				{
-					Route: "foo",
-				},
-				{
-					Route: "bar",
-				},
-			},
-		},
-		{
-			name: "plugins on combination of route and consumer",
-			args: args{
-				relations: foreignRelations{
-					Route:    []string{"foo", "bar"},
-					Consumer: []string{"foo", "bar"},
-				},
-			},
-			want: []rel{
-				{
-					Consumer: "foo",
-					Route:    "foo",
-				},
-				{
-					Consumer: "bar",
-					Route:    "foo",
-				},
-				{
-					Consumer: "foo",
-					Route:    "bar",
-				},
-				{
-					Consumer: "bar",
-					Route:    "bar",
-				},
-			},
-		},
-		{
-			name: "plugins on combination of service and consumer",
-			args: args{
-				relations: foreignRelations{
-					Service:  []string{"foo", "bar"},
-					Consumer: []string{"foo", "bar"},
-				},
-			},
-			want: []rel{
-				{
-					Consumer: "foo",
-					Service:  "foo",
-				},
-				{
-					Consumer: "bar",
-					Service:  "foo",
-				},
-				{
-					Consumer: "foo",
-					Service:  "bar",
-				},
-				{
-					Consumer: "bar",
-					Service:  "bar",
-				},
-			},
-		},
-		{
-			name: "plugins on combination of service,route and consumer",
-			args: args{
-				relations: foreignRelations{
-					Consumer: []string{"c1", "c2"},
-					Route:    []string{"r1", "r2"},
-					Service:  []string{"s1", "s2"},
-				},
-			},
-			want: []rel{
-				{
-					Consumer: "c1",
-					Service:  "s1",
-				},
-				{
-					Consumer: "c2",
-					Service:  "s1",
-				},
-				{
-					Consumer: "c1",
-					Service:  "s2",
-				},
-				{
-					Consumer: "c2",
-					Service:  "s2",
-				},
-				{
-					Consumer: "c1",
-					Route:    "r1",
-				},
-				{
-					Consumer: "c2",
-					Route:    "r1",
-				},
-				{
-					Consumer: "c1",
-					Route:    "r2",
-				},
-				{
-					Consumer: "c2",
-					Route:    "r2",
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := getCombinations(tt.args.relations); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("getCombinations() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_overrideRouteStripPath(t *testing.T) {
-	type args struct {
-		route *kong.Route
-		anns  map[string]string
-	}
-	tests := []struct {
-		name string
-		args args
-		want *kong.Route
-	}{
-		{},
-		{
-			name: "basic empty route",
-			args: args{
-				route: &kong.Route{},
-			},
-			want: &kong.Route{},
-		},
-		{
-			name: "set to false",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"configuration.konghq.com/strip-path": "false",
-				},
-			},
-			want: &kong.Route{
-				StripPath: kong.Bool(false),
-			},
-		},
-		{
-			name: "set to true and case insensitive",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"configuration.konghq.com/strip-path": "truE",
-				},
-			},
-			want: &kong.Route{
-				StripPath: kong.Bool(true),
-			},
-		},
-		{
-			name: "overrides any other value",
-			args: args{
-				route: &kong.Route{
-					StripPath: kong.Bool(false),
-				},
-				anns: map[string]string{
-					"configuration.konghq.com/strip-path": "truE",
-				},
-			},
-			want: &kong.Route{
-				StripPath: kong.Bool(true),
-			},
-		},
-		{
-			name: "random value",
-			args: args{
-				route: &kong.Route{
-					StripPath: kong.Bool(false),
-				},
-				anns: map[string]string{
-					"configuration.konghq.com/strip-path": "42",
-				},
-			},
-			want: &kong.Route{
-				StripPath: kong.Bool(false),
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			overrideRouteStripPath(tt.args.route, tt.args.anns)
-			if !reflect.DeepEqual(tt.args.route, tt.want) {
-				t.Errorf("overrideRouteStripPath() got = %v, want %v", tt.args.route, tt.want)
-			}
-		})
-	}
-}
-
-func Test_overrideServicePath(t *testing.T) {
-	type args struct {
-		service *kong.Service
-		anns    map[string]string
-	}
-	tests := []struct {
-		name string
-		args args
-		want *kong.Service
-	}{
-		{},
-		{
-			name: "basic empty service",
-			args: args{
-				service: &kong.Service{},
-			},
-			want: &kong.Service{},
-		},
-		{
-			name: "set to valid value",
-			args: args{
-				service: &kong.Service{},
-				anns: map[string]string{
-					"configuration.konghq.com/path": "/foo",
-				},
-			},
-			want: &kong.Service{
-				Path: kong.String("/foo"),
-			},
-		},
-		{
-			name: "does not set path if doesn't start with /",
-			args: args{
-				service: &kong.Service{},
-				anns: map[string]string{
-					"configuration.konghq.com/path": "foo",
-				},
-			},
-			want: &kong.Service{},
-		},
-		{
-			name: "overrides any other value",
-			args: args{
-				service: &kong.Service{
-					Path: kong.String("/foo"),
-				},
-				anns: map[string]string{
-					"configuration.konghq.com/path": "/bar",
-				},
-			},
-			want: &kong.Service{
-				Path: kong.String("/bar"),
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			overrideServicePath(tt.args.service, tt.args.anns)
-			if !reflect.DeepEqual(tt.args.service, tt.want) {
-				t.Errorf("overrideServicePath() got = %v, want %v", tt.args.service, tt.want)
-			}
-		})
-	}
-}
-
-func Test_overrideRouteHTTPSRedirectCode(t *testing.T) {
-	type args struct {
-		route *kong.Route
-		anns  map[string]string
-	}
-	tests := []struct {
-		name string
-		args args
-		want *kong.Route
-	}{
-		{},
-		{
-			name: "basic empty route",
-			args: args{
-				route: &kong.Route{},
-			},
-			want: &kong.Route{},
-		},
-		{
-			name: "basic sanity",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/https-redirect-status-code": "301",
-				},
-			},
-			want: &kong.Route{
-				HTTPSRedirectStatusCode: kong.Int(301),
-			},
-		},
-		{
-			name: "random integer value",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/https-redirect-status-code": "42",
-				},
-			},
-			want: &kong.Route{},
-		},
-		{
-			name: "random string",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/https-redirect-status-code": "foo",
-				},
-			},
-			want: &kong.Route{},
-		},
-		{
-			name: "force ssl annotation set to true and protocols is not set",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"ingress.kubernetes.io/force-ssl-redirect": "true",
-				},
-			},
-			want: &kong.Route{
-				HTTPSRedirectStatusCode: kong.Int(302),
-				Protocols:               []*string{kong.String("https")},
-			},
-		},
-		{
-			name: "force ssl annotation set to true and protocol is set to grpc",
-			args: args{
-				route: &kong.Route{
-					Protocols: []*string{kong.String("grpc")},
-				},
-				anns: map[string]string{
-					"ingress.kubernetes.io/force-ssl-redirect": "true",
-					"konghq.com/protocols":                     "grpc",
-				},
-			},
-			want: &kong.Route{
-				HTTPSRedirectStatusCode: kong.Int(302),
-				Protocols:               []*string{kong.String("grpcs")},
-			},
-		},
-		{
-			name: "force ssl annotation set to false",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"ingress.kubernetes.io/force-ssl-redirect": "false",
-				},
-			},
-			want: &kong.Route{},
-		},
-		{
-			name: "force ssl annotation set to true and HTTPS redirect code set to 307",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"ingress.kubernetes.io/force-ssl-redirect": "true",
-					"konghq.com/https-redirect-status-code":    "307",
-				},
-			},
-			want: &kong.Route{
-				HTTPSRedirectStatusCode: kong.Int(307),
-				Protocols:               []*string{kong.String("https")},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			overrideRouteHTTPSRedirectCode(tt.args.route, tt.args.anns)
-			if !reflect.DeepEqual(tt.args.route, tt.want) {
-				t.Errorf("overrideRouteHTTPSRedirectCode() got = %v, want %v", tt.args.route, tt.want)
-			}
-		})
-	}
-}
-
-func Test_overrideRoutePreserveHost(t *testing.T) {
-	type args struct {
-		route *kong.Route
-		anns  map[string]string
-	}
-	tests := []struct {
-		name string
-		args args
-		want *kong.Route
-	}{
-		{},
-		{
-			name: "basic empty route",
-			args: args{
-				route: &kong.Route{},
-			},
-			want: &kong.Route{},
-		},
-		{
-			name: "basic sanity",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/preserve-host": "true",
-				},
-			},
-			want: &kong.Route{
-				PreserveHost: kong.Bool(true),
-			},
-		},
-		{
-			name: "case insensitive",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/preserve-host": "faLSe",
-				},
-			},
-			want: &kong.Route{
-				PreserveHost: kong.Bool(false),
-			},
-		},
-		{
-			name: "random integer value",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/https-redirect-status-code": "42",
-				},
-			},
-			want: &kong.Route{},
-		},
-		{
-			name: "random string",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/https-redirect-status-code": "foo",
-				},
-			},
-			want: &kong.Route{},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			overrideRoutePreserveHost(tt.args.route, tt.args.anns)
-			if !reflect.DeepEqual(tt.args.route, tt.want) {
-				t.Errorf("overrideRoutePreserveHost() got = %v, want %v", tt.args.route, tt.want)
-			}
-		})
-	}
-}
-
-func Test_overrideRouteRegexPriority(t *testing.T) {
-	type args struct {
-		route *kong.Route
-		anns  map[string]string
-	}
-	tests := []struct {
-		name string
-		args args
-		want *kong.Route
-	}{
-		{},
-		{
-			name: "basic empty route",
-			args: args{
-				route: &kong.Route{},
-			},
-			want: &kong.Route{},
-		},
-		{
-			name: "basic sanity",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/regex-priority": "10",
-				},
-			},
-			want: &kong.Route{
-				RegexPriority: kong.Int(10),
-			},
-		},
-		{
-			name: "negative integer",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/regex-priority": "-10",
-				},
-			},
-			want: &kong.Route{
-				RegexPriority: kong.Int(-10),
-			},
-		},
-		{
-			name: "random float value",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/regex-priority": "42.42",
-				},
-			},
-			want: &kong.Route{},
-		},
-		{
-			name: "random string",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/regex-priority": "foo",
-				},
-			},
-			want: &kong.Route{},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			overrideRouteRegexPriority(tt.args.route, tt.args.anns)
-			if !reflect.DeepEqual(tt.args.route, tt.want) {
-				t.Errorf("overrideRouteRegexPriority() got = %v, want %v", tt.args.route, tt.want)
-			}
-		})
-	}
-}
-
-func Test_overrideRouteMethods(t *testing.T) {
-	type args struct {
-		route *kong.Route
-		anns  map[string]string
-	}
-	tests := []struct {
-		name string
-		args args
-		want *kong.Route
-	}{
-		{},
-		{
-			name: "basic empty route",
-			args: args{
-				route: &kong.Route{},
-			},
-			want: &kong.Route{},
-		},
-		{
-			name: "basic sanity",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/methods": "POST,GET",
-				},
-			},
-			want: &kong.Route{
-				Methods: kong.StringSlice("POST", "GET"),
-			},
-		},
-		{
-			name: "non-string",
-			args: args{
-				route: &kong.Route{},
-				anns: map[string]string{
-					"konghq.com/methods": "-10,GET",
-				},
-			},
-			want: &kong.Route{},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			overrideRouteMethods(logrus.New(), tt.args.route, tt.args.anns)
-			if !reflect.DeepEqual(tt.args.route, tt.want) {
-				t.Errorf("overrideRouteMethods() got = %v, want %v", tt.args.route, tt.want)
 			}
 		})
 	}
